Refactor XMLParserTest
